o/servicestate: move handlers tests to quota_handlers_test.go file instead
These tests are now exclusively about the handlers, so they should live in the
handlers specific test file.
